+++ trainingeditor/src/main/java/com/dr/trainup/trainingeditor/ui/vm/TrainingEditorViewModel.kt
@@ -80,7 +80,7 @@ class TrainingEditorViewModel @Inject constructor(
                 Station(
                     stationID,
                     stationName,
-                    weight.toInt(),
+                    weight.toFloat(),
                     weightUnit,
                     repeats.toInt(),
                     seatPosition
